:art: Componentize the details splash area
