+++ src/app/cube/details/details.component.scss
@@ -7,104 +7,6 @@ $content-width: calc(#{$max-width} - #{$sidebar-width});
   width: 100%;
   padding-bottom: 20px;
 
-  &.nanomodule {
-    .type {
-      background: $nanomodule-color;
-    }
-  }
-
-  &.micromodule {
-    .type {
-      background: $micromodule-color;
-    }
-  }
-
-  &.module {
-    .type {
-      background: $module-color;
-    }
-  }
-
-  &.unit {
-    .type {
-      background: $unit-color;
-    }
-  }
-
-  &.course {
-    .type {
-      background: $course-color;
-    }
-  }
-
-  .splash {
-    background: $blue-gradient;
-    padding: 50px 0;
-    position: relative;
-
-    &:after {
-      content: '';
-      background: white;
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      opacity: 0.03;
-      mask: $cube-pattern repeat;
-    }
-
-    .inner {
-      max-width: $max-width;
-      margin: auto;
-      padding: 20px;
-      box-sizing: border-box;
-      z-index: 1;
-
-      .header {
-        max-width: $content-width;
-        padding-right: 20px;
-      }
-
-      .name {
-        font-size: 40px;
-        color: white;
-        text-shadow: 0 0 rgba(0, 0, 0, 0.2);
-      }
-
-      .author,
-      .downloads {
-        color: white;
-        font-size: 18px;
-        font-weight: 300;
-      }
-
-      .type {
-        margin-top: 5px;
-        padding: 3px 8px;
-        border-radius: 3px;
-        display: inline-block;
-        position: relative;
-        color: white;
-        text-transform: capitalize;
-        font-size: 14px;
-        font-weight: 300;
-
-        &:before {
-          content: '';
-          border-radius: 50%;
-          width: 10px;
-          height: 10px;
-          left: 0;
-          top: 0;
-          bottom: 0;
-          margin: auto;
-          position: absolute;
-        }
-      }
-    }
-  }
-
   .flex {
     display: flex;
     flex-direction: row;
