+++ src/app/cube/details/included-in/included-in.component.scss
@@ -9,6 +9,7 @@
     padding: 8px;
     font-size: 16px;
     background-color: white;
+
     .lo-length {
       text-transform: capitalize;
     }
