+++ src/app/cube/details/splash/details-splash.component.html
@@ -1,0 +1,16 @@
+<div class="splash">
+  <div class="inner">
+    <div class="header">
+      <div class="name">{{ learningObject.name }}</div>
+      <div class="author">Created by
+        <span [routerLink]="['/users', learningObject.author.username]" class="link">{{learningObject.author.name + " - " + learningObject.author.organization | titlecase}}</span>, updated {{ learningObject.date
+        | date:'shortDate' }}</div>
+      <div *ngIf="learningObject.contributors" class="author">Contributors:
+        <span *ngFor="let contributor of contributorsList; let i = index;" class="link" [routerLink]="['/users', contributor.username]">{{ contributor.name + " - " + contributor.organization | titlecase }}
+          <span *ngIf="i !== contributorsList.length -1">, </span>
+        </span>
+      </div>
+      <div [ngClass]="learningObject.length" class="type">{{ learningObject.length }}</div>
+    </div>
+  </div>
+</div>
